product in cart is now a separate component
